+++ packages/asc-ui/src/components/Label/LabelContext.ts
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const LabelContext = React.createContext({
+  active: false,
+  setActive: () => {},
+} as { active: boolean; setActive: Function })
+export default LabelContext